requirements.txt
numpy>=1.19.2
pandas>=1.2.0
matplotlib>=3.3.2
seaborn>=0.11.0
tensorflow>=2.4.0
scikit-learn>=0.24.0
opencv-python>=4.4.0
kaggle>=1.5.12
tqdm>=4.50.0
arabic-reshaper>=2.1.3
python-bidi>=0.4.2 